Move submission print-view loader and types to utils/submission
